# Vectorize some Guid APIs (ctor, TryWriteBytes)