--- a/src/System.Private.CoreLib/shared/System/Guid.cs
+++ b/src/System.Private.CoreLib/shared/System/Guid.cs
@@ -786,7 +786,7 @@ namespace System
         // Returns whether bytes are sucessfully written to given span.
         public bool TryWriteBytes(Span<byte> destination)
         {
-            if ((uint)destination.Length < 16)
+            if (destination.Length < 16)
                 return false;
 
             if (BitConverter.IsLittleEndian)
@@ -795,6 +795,7 @@ namespace System
             }
             else
             {
+                destination[15] = _k;
                 destination[0] = (byte)(_a);
                 destination[1] = (byte)(_a >> 8);
                 destination[2] = (byte)(_a >> 16);
@@ -810,7 +811,6 @@ namespace System
                 destination[12] = _h;
                 destination[13] = _i;
                 destination[14] = _j;
-                destination[15] = _k;
             }
             return true;
         }