--- a/src/System.Private.CoreLib/shared/System/Guid.cs
+++ b/src/System.Private.CoreLib/shared/System/Guid.cs
@@ -786,16 +786,14 @@ namespace System
         // Returns whether bytes are sucessfully written to given span.
         public bool TryWriteBytes(Span<byte> destination)
         {
-            if (destination.Length < 16)
-                return false;
-
             if (BitConverter.IsLittleEndian)
             {
-                MemoryMarshal.Write(destination, ref this);
+                return MemoryMarshal.TryWrite(destination, ref this);
             }
-            else
+
+            // slower path for BigEndian
+            if ((uint)destination.Length >= 16)
             {
-                destination[15] = _k;
                 destination[0] = (byte)(_a);
                 destination[1] = (byte)(_a >> 8);
                 destination[2] = (byte)(_a >> 16);
@@ -811,8 +809,10 @@ namespace System
                 destination[12] = _h;
                 destination[13] = _i;
                 destination[14] = _j;
+                destination[15] = _k;
+                return true;
             }
-            return true;
+            return false;
         }
 
         // Returns the guid in "registry" format.