1. `Guid.TryWriteBytes(Span<byte>)`:
```csharp
[Benchmark]
[ArgumentsSource(nameof(TestData))]
public Span<byte> TryWriteBytes(Guid id)
{
    Span<byte> s = new byte[20];
    id.TryWriteBytes(s);
    return s;
}
```
|            Method |     Mean |     Error |    StdDev |
|------------------ |---------:|----------:|----------:|
| TryWriteBytes_new | 4.960 ns | 0.0818 ns | 0.0292 ns |
| TryWriteBytes_old | 8.541 ns | 0.0362 ns | 0.0129 ns |
</br>

2. `new Guid(byte[])` and `new Guid(ReadOnlySpan<byte>)`:
```csharp
[Benchmark]
[ArgumentsSource(nameof(TestData))]
public Guid GuidCtor(byte[] data)
{
    return new Guid(data);
}
```

|       Method |     Mean |     Error |    StdDev |
|------------- |---------:|----------:|----------:|
| GuidCtor_new | 2.615 ns | 0.0275 ns | 0.0098 ns |
| GuidCtor_old | 7.483 ns | 0.0146 ns | 0.0052 ns |
 
</br>

3. `Guid.ToByteArray()`:
```csharp
[Benchmark]
[ArgumentsSource(nameof(TestData))]
public byte[] ToByteArray(Guid id)
{
    return id.ToByteArray();
}
```
|          Method |     Mean |     Error |    StdDev |
|---------------- |---------:|----------:|----------:|
| ToByteArray_new | 3.569 ns | 0.0192 ns | 0.0068 ns |
| ToByteArray_old | 6.265 ns | 0.0776 ns | 0.0202 ns |

 
</br>

Environment:
``` ini

BenchmarkDotNet=v0.11.3, OS=Windows 10.0.17134.407 (1803/April2018Update/Redstone4)
Intel Core i7-8700K CPU 3.70GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
Frequency=3609372 Hz, Resolution=277.0565 ns, Timer=TSC
.NET Core SDK=3.0.100-preview-009812
  [Host]     : .NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103), 64bit RyuJIT
  Job-CIYVXD : .NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103), 64bit RyuJIT

IterationCount=6  WarmupCount=6  

```