@jkotas 2018-12-03
> https://gist.github.com/EgorBo/75348d5f194cbe91342d1b77d73b6c04

I has AggressiveInlining on everything. The actual implementation does not have it. Are the results representative?