@EgorBo 2018-12-03
@jkotas oh, indeed, let me re-run the benchmark without it. Initially I wanted to replace duplicated code to compare guids with `Equals` with [AggressiveInline] e.g. https://github.com/dotnet/coreclr/blob/master/src/System.Private.CoreLib/shared/System/Guid.cs#L833-L845 (+ operator ==/!=)