@EgorBo 2018-12-03
hm.. without `` results a little bit different:
![image](https://user-images.githubusercontent.com/523221/49375559-b3efa100-f715-11e8-8e2c-9748e9c72503.png)
comparing Guids as two ulongs is the most efficient way now (tried several times on a clean env).