@EgorBo 2018-12-03
Turns out `WriteByteHelper` can be optimized with SSE2 (it makes it twice faster in my cases). But it also depends on alignment. PS: some other similar assignments can be optimized like this too here.