@EgorBo 2018-12-03
@jkotas I've removed `Equals` fix (will do that in a separate PR with other places).
I'll rewrite the PR description in an hour and attach better benchmark for what we've done now 🙂