@jkotas 2018-12-03
Could you please also add a short comment like `// Hoist bounds checks` next to all of the out-of-order writes so that somebody looking at this will get a hint that it is done intentionally this way?